For each file, store a referencing translation unit
Parsing header files on their own does often not allow finding definitions.
When jumping to a location we keep track of the referencing translation
unit. If we then try to find a definition of something contained in the
header file, we refer back to the original translation unit that included
the header file.